+++ src/client/index.js
@@ -1,0 +1,3 @@
+import YFClient from './node_client'
+
+export default YFClient