+++ src/client/node_client.js
@@ -1,10 +1,10 @@
 import _ from 'lodash'
-import fetch from 'node-fetch'
-import md5 from 'md5'
+import axios from 'axios'
+import { md5 } from './kit'
 
-function sign(args){
+const sign = (args) =>{
   return md5(
-      _.map(args, function(val, key){
+      _.map(args, (val, key) => {
         if(_.isObject(val)){
           val = JSON.stringify(val)
         }
@@ -15,12 +15,7 @@ function sign(args){
 
 const defaultOptions = {
   v: '0.0.1',
-  endpoint: 'http://api.yunplus.io/api',
-  upload: 'http://api.yunplus.io/upload',
-  fields: {
-    createAt: { column: 'createAt', type: 'bigint', },
-    updateAt: { column: 'updateAt', type: 'bigint', }
-  }
+  domain: 'http://localhost:9999', // an domain url or ip:port ex: http://localhost:9999
 }
 
 class YFClient {
@@ -33,40 +28,18 @@ class YFClient {
 
   static init(options){
     YFClient._options = _.assign(defaultOptions, options)
-    YFClient._options.upload = YFClient._options.endpoint.replace(/(\/api|\/api\/)$/, '/upload')
-    YFClient._options.ping = YFClient._options.endpoint.replace(/(\/api|\/api\/)$/, '/ping')
+    YFClient._options.endpoint = YFClient._options.domain + '/api'
+    YFClient._options.ping = YFClient._options.domain + '/ping'
   }
 
   static ping(){
     return new Promise( (resolve, reject) => {
-      fetch(YFClient._options.ping)
+      axios.get(YFClient._options.ping)
         .then((rsp) => {
-          return rsp.json()
+          return rsp.data
         })
         .then((json) => {
           resolve(json)
-        })
-        .catch((err) => {
-          reject(err)
-        })
-    })
-  }
-
-  upload(data){
-    return new Promise( (resolve, reject) => {
-      fetch(YFClient._options.upload, {
-        method: 'POST',
-        body: data
-      })
-        .then((rsp) => {
-          return rsp.json()
-        })
-        .then((json) => {
-          if(json.errno === 0){
-            resolve(json.data)
-          }else{
-            reject(json)
-          }
         })
         .catch((err) => {
           reject(err)
@@ -89,15 +62,9 @@ class YFClient {
     delete inputData.masterKey
 
     return new Promise( (resolve, reject) => {
-      fetch(YFClient._options.endpoint, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json; charset=UTF-8'
-        },
-        body: JSON.stringify (inputData)
-      })
+      axios.post(YFClient._options.endpoint, inputData)
         .then((rsp) => {
-          return rsp.json()
+          return rsp.data
         })
         .then((json) => {
           if(json.errno === 0){